+++ cmd/curator/main.go
@@ -21,7 +21,7 @@ import (
 )
 
 const (
-	version = "0.17.1"
+	version = "0.17.2"
 )
 
 func main() {